Alle errors & warnings opgelost + gitignore toegevoegd

+++ RoguelikeEndAssesment/RoguelikeEndAssesment/RogueLike/Utils/Random.h
@@ -12,7 +12,7 @@ public:
 	static T GetRandom(const int min, const int max)
 	{
 		std::default_random_engine generator;
-		generator.seed(time(0));
+		generator.seed((unsigned int)time(0));
 		std::uniform_int_distribution<T> distribution1(min, max);
 		return distribution1(generator);
 	}
